backend/requirements.txt
fastapi>=0.116.1
uvicorn>=0.34.0
pydantic>=2.11.0
pydantic-settings>=2.10.1
python-dotenv>=1.1.1
google-generativeai>=0.8.5
google-genai>=0.1.0
chromadb>=0.5.0
pypdf
PyYAML
httpx
youtube-transcript-api
google-api-python-client